==> 0_ecog_cog_change_filepaths.R <==
# Author: Natalie Gradwohl
# Creation Date: 11/16/2023
# Purpose: Save paths to folders/files for easy access to datasets/outputs

rm(list = ls())

path_to_box <- "C:/Users/Natalie Gradwohl/Box/"

path_to_raw_data <-
  paste0(
    path_to_box,
    "KHANDLE_coded_data/Raw_data_tables/All_Wave/transfer15_2024Apr19/khandle_all_waves_20240419.sas7bdat"
  )

# path_to_raw_data <- r"(DRIVE:\...\khandle_all_waves_20230912.sas7bdat)"
# #<- FOR KAISER USE: add raw data filepath. Please use the PHI KHANDLE dataset from September 2023
# 
# path_to_output <- r"(DRIVE:\...\test_output\)"
# #<- FOR KAISER USE: add outputs folder. Be sure to put a a backslash at the end of the directory
